Bug 1510490 - Measure shader cache memory usage and remove total_gpu_bytes. r=mattwoodrow
The latter causes confusion in the memory reports because it gets summed
up and thus effectively doubles the reported texture memory usage. I've
decided it's best to drop, and so might as well do that while we're
already messing around with the memory reports and the associated
boilerplate.

Depends on D13439

Differential Revision: https://phabricator.services.mozilla.com/D13440

--HG--
extra : moz-landing-system : lando