Bug 1535146 - Use a whitelist to decide which shaders to load from disk on startup. r=bholley
Webrender caches the program binaries of shaders used within the first
ten frames, so that on next startup it can load them from disk rather
than having to recompile them.

Previously it would load all binaries found in the disk cache on
startup, and when saving to the cache it would delete any existing
binaries that weren't used.

This changes it so that unused binaries are not deleted. The disk
space this requires is insignificant, but as the cache grows loading
all the shaders on startup can get expensive. To solve that we write a
whitelist of the shaders used during startup, and only load those
during the next startup.

Differential Revision: https://phabricator.services.mozilla.com/D33954

--HG--
extra : moz-landing-system : lando