Bug 1614679 - Cache compiled shaders to disk even after startup. r=jrmuizel
Previously we only saved shaders to disk on the tenth frame, meaning
shaders compiled afterwards would not be cached and would need to be
recompiled in future runs of the application. This change makes it so
that we cache shaders to disk regardless of which frame they are
compiled in.

We continue to treat only the shaders used within the first ten frames
as the "startup shaders", meaning only those ones will be loaded on
next startup. Caching as many other shaders as possible is still
beneficial, however, as they are loaded on-demand.

Differential Revision: https://phabricator.services.mozilla.com/D62748

--HG--
extra : moz-landing-system : lando